move footer to own layout with styles
