+++ src/styles/main.scss
@@ -11,7 +11,8 @@
 @import '../components/ImageWithCredit/ImageWithCredits';
 @import '../components/NavBar/NavBar';
 
-/* Pages */
+/* Layout */
+@import '../layout/Footer/Footer';
 
 /* Utils */
 @import './link';
